Refactored bulk of transaction handling code to its own module
